+++ DICUI.Test/DICUI.Test.csproj
@@ -63,7 +63,6 @@
     <Compile Include="UI\AllowedSpeedsTest.cs" />
     <Compile Include="Utilities\DumpEnvironmentTest.cs" />
     <Compile Include="ResultTest.cs" />
-    <Compile Include="TasksTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utilities\DriveTest.cs" />
     <Compile Include="Utilities\ConvertersTest.cs" />
