--- DICUI.Test/TasksTest.cs
@@ -1,30 +1,0 @@
-﻿using DICUI.Data;
-using DICUI.Utilities;
-using Xunit;
-
-namespace DICUI.Test
-{
-    public class TasksTest
-    {
-        [Fact]
-        public void EjectDiscTest()
-        {
-            // TODO: Implement
-            Assert.True(true);
-        }
-
-        [Fact]
-        public void CancelDumpingTest()
-        {
-            // TODO: Implement
-            Assert.True(true);
-        }
-
-        [Fact]
-        public void StartDumpingTest()
-        {
-            // TODO: Implement
-            Assert.True(true);
-        }
-    }
-}
