+++ DICUI.Test/Utilities/DumpEnvironmentTest.cs
@@ -123,5 +123,26 @@ namespace DICUI.Test
             // TODO: Implement
             Assert.True(true);
         }
+
+        [Fact]
+        public void EjectDiscTest()
+        {
+            // TODO: Implement
+            Assert.True(true);
+        }
+
+        [Fact]
+        public void CancelDumpingTest()
+        {
+            // TODO: Implement
+            Assert.True(true);
+        }
+
+        [Fact]
+        public void StartDumpingTest()
+        {
+            // TODO: Implement
+            Assert.True(true);
+        }
     }
 }
