+++ DICUI/DICUI.csproj
@@ -94,12 +94,12 @@
     <Compile Include="Data\Constants.cs" />
     <Compile Include="External\EVORE.cs" />
     <Compile Include="Options.cs" />
-    <Compile Include="Tasks.cs" />
     <Compile Include="External\ProtectionFind.cs" />
     <Compile Include="UI\KnownSystemComboBoxItem.cs" />
     <Compile Include="UI\AllowedSpeeds.cs" />
     <Compile Include="Utilities\DumpEnvironment.cs" />
     <Compile Include="Utilities\Extensions.cs" />
+    <Compile Include="Utilities\Result.cs" />
     <Compile Include="Utilities\Validators.cs" />
     <Compile Include="Utilities\Converters.cs" />
     <Compile Include="OptionsWindow.xaml.cs">
