+++ DICUI/Utilities/Result.cs
@@ -1,8 +1,4 @@
-﻿using System.Linq;
-using System.Threading.Tasks;
-using DICUI.External;
-
-namespace DICUI
+﻿namespace DICUI.Utilities
 {
     /// <summary>
     /// Generic success/failure result object, with optional message
@@ -27,29 +23,5 @@ namespace DICUI
         public static Result Failure(string message, params object[] args) => new Result(false, string.Format(message, args));
 
         public static implicit operator bool(Result result) => result.success;
-    }
-
-    /// <summary>
-    /// Class containing dumping tasks
-    /// </summary>
-    public class Tasks
-    {
-        /// <summary>
-        /// Run protection scan on a given dump environment
-        /// </summary>
-        /// <param name="env">DumpEnvirionment containing all required information</param>
-        /// <returns>Copy protection detected in the envirionment, if any</returns>
-        public static async Task<string> RunProtectionScan(string path)
-        {
-            var found = await Task.Run(() =>
-            {
-                return ProtectionFind.Scan(path);
-            });
-
-            if (found == null)
-                return "None found";
-
-            return string.Join("\n", found.Select(kvp => kvp.Key + ": " + kvp.Value).ToArray());
-        }
     }
 }
